+++ packages/website/sidebars/sidebar.base.js
@@ -23,6 +23,7 @@ module.exports = {
           collapsible: false,
           type: 'category',
           items: [
+            'linting/troubleshooting/performance-troubleshooting',
             'linting/troubleshooting/formatting',
             'linting/troubleshooting/tslint',
           ],
@@ -35,30 +36,62 @@ module.exports = {
       label: 'Getting Started',
       type: 'category',
     },
+    'custom-rules',
     {
-      type: 'category',
-      label: 'Development',
-      collapsible: false,
       items: [
-        {
-          label: 'Architecture',
-          type: 'category',
-          collapsible: false,
-          items: [
-            'development/architecture/asts',
-            'development/architecture/packages',
-          ],
-        },
-        'development/custom-rules',
+        'architecture/eslint-plugin',
+        'architecture/eslint-plugin-tslint',
+        'architecture/parser',
+        'architecture/scope-manager',
+        'architecture/typescript-estree',
+        'architecture/utils',
       ],
+      label: 'Architecture',
+      link: {
+        id: 'architecture',
+        type: 'doc',
+      },
+      type: 'category',
     },
     {
-      collapsible: false,
       items: [
-        'maintenance/issues',
+        'contributing/discussions',
+        'contributing/issues',
+        'contributing/local-development',
+        'contributing/pull-requests',
+      ],
+      label: 'Contributing',
+      link: {
+        id: 'contributing',
+        type: 'doc',
+      },
+      type: 'category',
+    },
+    {
+      items: [
+        'maintenance/branding',
+        {
+          collapsible: false,
+          items: ['maintenance/issues/rule-deprecations'],
+          label: 'Issues',
+          link: {
+            id: 'maintenance/issues',
+            type: 'doc',
+          },
+          type: 'category',
+        },
         'maintenance/pull-requests',
         'maintenance/releases',
-        'maintenance/versioning',
+        {
+          collapsible: false,
+          items: ['maintenance/versioning/dependant-version-upgrades'],
+          label: 'Versioning',
+          link: {
+            id: 'maintenance/versioning',
+            type: 'doc',
+          },
+          type: 'category',
+        },
       ],
       label: 'Maintenance',
       link: {