+++ packages/website/src/components/FinancialContributors/Sponsor.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 
 import styles from './styles.module.css';
-import type { SponsorData, SponsorIncludeOptions } from './types';
+import type { SponsorData } from './types';
 
 interface SponsorProps {
-  include?: SponsorIncludeOptions;
+  includeName?: boolean;
   sponsor: SponsorData;
 }
 
-export function Sponsor({ include = {}, sponsor }: SponsorProps): JSX.Element {
+export function Sponsor({ includeName, sponsor }: SponsorProps): JSX.Element {
   let children = <img alt={`${sponsor.name} logo`} src={sponsor.image} />;
 
-  if (include.name) {
+  if (includeName) {
     children = (
       <>
         {children}
@@ -20,19 +20,15 @@ export function Sponsor({ include = {}, sponsor }: SponsorProps): JSX.Element {
     );
   }
 
-  if (include.link) {
-    children = (
-      <a
-        className={styles.sponsorLink}
-        href={sponsor.website ?? undefined}
-        title={sponsor.name}
-        target="_blank"
-        rel="noopener sponsored"
-      >
-        {children}
-      </a>
-    );
-  }
-
-  return children;
+  return (
+    <a
+      className={styles.sponsorLink}
+      href={sponsor.website ?? undefined}
+      title={sponsor.name}
+      target="_blank"
+      rel="noopener sponsored"
+    >
+      {children}
+    </a>
+  );
 }